refactor(sentry): extract sanitization from main.tsx and dedupe similar e2e code
