+++ tools/computer-use-e2e/peekaboo-workflow-contract-self-test.mjs
@@ -23,6 +23,9 @@ const frontendAppPath = path.join(repoRoot, 'apps/native/src/App.tsx');
 const frontendWidgetPath = path.join(repoRoot, 'apps/native/src/components/widget/widget.tsx');
 const frontendEditorPanelPath = path.join(repoRoot, 'apps/native/src/components/widget/overlays/editor-panel.tsx');
 const frontendBootDiagnosticsPath = path.join(repoRoot, 'apps/native/src/lib/e2e-boot-diagnostics.ts');
+const frontendSentryInitPath = path.join(repoRoot, 'apps/native/src/lib/sentry/init.ts');
+const frontendSentrySanitizePath = path.join(repoRoot, 'apps/native/src/lib/sentry/sanitize.ts');
+const frontendStartupFallbackPath = path.join(repoRoot, 'apps/native/src/components/StartupFallback.tsx');
 const tauriApiPath = path.join(repoRoot, 'apps/native/src/tauri-api.ts');
 const workflow = readFileSync(workflowPath, 'utf8');
 const productProof = readFileSync(productProofPath, 'utf8');
@@ -41,6 +44,9 @@ const frontendApp = readFileSync(frontendAppPath, 'utf8');
 const frontendWidget = readFileSync(frontendWidgetPath, 'utf8');
 const frontendEditorPanel = readFileSync(frontendEditorPanelPath, 'utf8');
 const frontendBootDiagnostics = readFileSync(frontendBootDiagnosticsPath, 'utf8');
+const frontendSentryInit = readFileSync(frontendSentryInitPath, 'utf8');
+const frontendSentrySanitize = readFileSync(frontendSentrySanitizePath, 'utf8');
+const frontendStartupFallback = readFileSync(frontendStartupFallbackPath, 'utf8');
 const tauriApi = readFileSync(tauriApiPath, 'utf8');
 
 function section(startPattern, endPattern = null) {
@@ -224,27 +230,33 @@ assert.match(frontendBootDiagnostics, /let bootStageCleared = false[\s\S]*export
 assert.match(frontendBootDiagnostics, /function markNativeBootStage\(stage: string\)[\s\S]*darwinAPI\.debug\.markBootStage\(stage, Date\.now\(\)\)[\s\S]*markNativeBootStage\(normalizedStage\)/, 'Frontend boot diagnostics must mirror boot stages to the native debug command before the event-loop stall boundary');
 assert.match(frontendBootDiagnostics, /export function clearBootStage[\s\S]*document\.title = APP_TITLE[\s\S]*nixmac:e2e-boot-stage", "mounted"/, 'Frontend boot diagnostics must restore the normal window title after mount');
 assert.match(frontendBootDiagnostics, /export function clearBootStage[\s\S]*markNativeBootStage\("mounted"\)/, 'Frontend boot diagnostics must clear the native title marker after App mount');
-assert.match(frontendBootDiagnostics, /export function sanitizeE2eDiagnosticText[\s\S]*EMAIL_PATTERN[\s\S]*BEARER_TOKEN_PATTERN[\s\S]*OPENAI_TOKEN_PATTERN[\s\S]*HOME_DIR_PATH_PATTERN/, 'E2E DOM diagnostics must sanitize secret-shaped text before persisting report artifacts');
+assert.match(frontendSentrySanitize, /export function sanitizeDiagnosticText[\s\S]*sanitizeString/, 'E2E DOM diagnostics text sanitizer must be exported from the shared sanitize module');
+assert.match(frontendSentrySanitize, /EMAIL_PATTERN[\s\S]*BEARER_TOKEN_PATTERN[\s\S]*OPENAI_TOKEN_PATTERN[\s\S]*HOME_DIR_PATH_PATTERN/, 'Shared sanitize module must apply the secret-shaped text patterns used by both Sentry events and E2E diagnostics');
+assert.match(frontendBootDiagnostics, /import \{ sanitizeDiagnosticText \} from "@\/lib\/sentry\/sanitize"/, 'E2E boot diagnostics must consume sanitization from the shared sentry/sanitize module rather than a duplicate regex set');
 assert.match(frontendBootDiagnostics, /export function recordE2eDomSnapshot[\s\S]*storagePrefix[\s\S]*nixmac:e2e-dom-snapshot[\s\S]*document\.documentElement\.dataset\.nixmacE2eDomSnapshot[\s\S]*`\$\{storagePrefix\}:last`[\s\S]*E2E DOM snapshot \$\{label\} summary[\s\S]*E2E DOM snapshot \$\{label\} text[\s\S]*E2E DOM snapshot \$\{label\} html/, 'E2E DOM diagnostics must persist bounded snapshots through both out-of-band DOM/localStorage state and breadcrumb artifacts');
 assert.match(frontendBootDiagnostics, /export function scheduleE2eDomSnapshots[\s\S]*count = 5[\s\S]*intervalMs = 2_000[\s\S]*emitted < count/, 'E2E DOM diagnostics must schedule a bounded post-mount snapshot series and self-stop');
-assert.match(frontendMain, /PREFS_BOOT_TIMEOUT_MS = 8000[\s\S]*ui_get_prefs invoke start[\s\S]*success after timeout[\s\S]*Promise\.race\(\[prefsPromise, timeoutPromise\]\)/, 'Frontend boot must log prefs IPC progress with clear after-timeout labels');
+assert.match(frontendSentryInit, /PREFS_BOOT_TIMEOUT_MS = 8000[\s\S]*ui_get_prefs invoke start[\s\S]*success after timeout[\s\S]*Promise\.race\(\[prefsPromise, timeoutPromise\]\)/, 'Sentry init module must log prefs IPC progress with clear after-timeout labels');
 assert.match(frontendMain, /markBootStage\("main-loaded"\)[\s\S]*markBootStage\("root-found"\)[\s\S]*markBootStage\("react-render-start"\)[\s\S]*markBootStage\("react-render-scheduled"\)/, 'Frontend boot must synchronously mark module, root, and render-scheduling stages');
 assert.match(frontendApp, /markBootStage\("app-render"\)[\s\S]*markBootStage\("app-effect"\)[\s\S]*clearBootStage\(\)/, 'App must synchronously mark render/effect stages and clear the E2E title marker after mount');
 assert.match(frontendWidget, /markBootStage\("darwin-widget-render"\)/, 'DarwinWidget must mark when the product widget render body is reached');
 assert.match(frontendEditorPanel, /const LazyNixEditor = lazy\(async \(\) => \{[\s\S]*import\("@\/components\/kibo-ui\/nix-editor"\)[\s\S]*default: module\.NixEditor/, 'EditorPanel must lazy-load the Monaco-backed Nix editor only when a file is opened');
 assert.doesNotMatch(frontendEditorPanel, /import \{ NixEditor \}/, 'EditorPanel must not import the Monaco-backed editor in the first app boot bundle');
 assert.match(frontendMain, /if \(E2E_MODE\) \{[\s\S]*setInterval\(\(\) => \{[\s\S]*boot heartbeat[\s\S]*boot heartbeat upper bound reached[\s\S]*stopBootHeartbeat[\s\S]*boot heartbeat stopped[\s\S]*nixmac:app-mounted/, 'Frontend boot must emit bounded E2E-only heartbeat breadcrumbs until App mounted and record when the bound is reached');
-assert.match(frontendMain, /E2E_MODE = import\.meta\.env\.VITE_NIXMAC_E2E_MODE === "true"[\s\S]*Sentry init skipped for E2E boot[\s\S]*return;/, 'Frontend boot must use the build-time E2E mode flag to skip boot-time Sentry prefs IPC without adding another IPC gate');
+assert.match(frontendMain, /E2E_MODE = import\.meta\.env\.VITE_NIXMAC_E2E_MODE === "true"/, 'Frontend main must derive E2E_MODE from the build-time Vite flag');
+assert.match(frontendSentryInit, /E2E_MODE = import\.meta\.env\.VITE_NIXMAC_E2E_MODE === "true"[\s\S]*Sentry init skipped for E2E boot[\s\S]*return;/, 'Sentry init module must use the build-time E2E mode flag to skip boot-time Sentry prefs IPC without adding another IPC gate');
 assert.match(frontendMain, /E2E_APP_MOUNT_RELOAD_TIMEOUT_MS = 12000[\s\S]*E2E_APP_MOUNT_RELOAD_KEY[\s\S]*E2E app-mounted watchdog reloading[\s\S]*window\.location\.reload\(\)/, 'Frontend E2E boot must request one reload when the page loads but App never mounts');
 assert.match(frontendMain, /scheduleE2eDomSnapshots\("post-mount"\)[\s\S]*recordE2eDomSnapshot\("app-mounted-watchdog-before-reload"[\s\S]*nixmac:e2e-dom-snapshot:watchdog-pre-reload[\s\S]*window\.setTimeout\(\(\) => \{[\s\S]*window\.location\.reload\(\)[\s\S]*250/, 'Frontend E2E boot must capture post-mount DOM snapshots and a durable watchdog snapshot before forced reload');
 assert.match(frontendMain, /const renderApp = \(\) => \{[\s\S]*React render start[\s\S]*Sentry\.ErrorBoundary[\s\S]*<App \/>[\s\S]*React render scheduled/, 'Frontend boot must render the app immediately with an error boundary');
-assert.match(frontendMain, /startSentryInitOnce[\s\S]*render-error[\s\S]*render-fatal[\s\S]*Sentry init start requested[\s\S]*scheduleAfterPostMountFrame[\s\S]*initializeSentryAfterPostMountFrame\(\)/, 'Frontend boot must start preference-backed Sentry initialization only after App mounted, render error, render fatal, or the mount-timeout fallback requests it');
-assert.match(frontendMain, /captureRenderErrorAfterSentryInit[\s\S]*startSentryInitOnce\(reason\)[\s\S]*Sentry\.captureException\(error\)/, 'Frontend boot must explicitly capture render errors after Sentry init is requested');
-assert.match(frontendMain, /SENTRY_MOUNT_TIMEOUT_MS = 5000[\s\S]*window\.setTimeout\(\(\) => \{[\s\S]*startSentryInitOnce\("mount-timeout"\)[\s\S]*SENTRY_MOUNT_TIMEOUT_MS/, 'Frontend boot must retain a production mount-timeout Sentry fallback for failed-render sessions without a long early-boot observability gap');
-assertOrder(frontendMain, 'window.addEventListener(\n  "nixmac:app-mounted"', 'renderApp();', 'Frontend boot must register the app-mounted listener before rendering the app');
+assert.match(frontendSentryInit, /startSentryInitOnce[\s\S]*render-error[\s\S]*render-fatal[\s\S]*Sentry init start requested[\s\S]*scheduleAfterPostMountFrame[\s\S]*initializeSentryAfterPostMountFrame\(\)/, 'Sentry init module must start preference-backed Sentry initialization only after App mounted, render error, render fatal, or the mount-timeout fallback requests it');
+assert.match(frontendSentryInit, /export function captureRenderError[\s\S]*startSentryInitOnce\(reason\)[\s\S]*Sentry\.captureException\(error\)/, 'Sentry init module must export captureRenderError that lazily initializes Sentry and then captures the render error');
+assert.match(frontendSentryInit, /SENTRY_MOUNT_TIMEOUT_MS = 5000[\s\S]*window\.setTimeout\(\(\) => \{[\s\S]*startSentryInitOnce\("mount-timeout"\)[\s\S]*SENTRY_MOUNT_TIMEOUT_MS/, 'Sentry init module must retain a production mount-timeout Sentry fallback for failed-render sessions without a long early-boot observability gap');
+assert.match(frontendSentryInit, /export function attachSentry[\s\S]*window\.addEventListener\(\s*"nixmac:app-mounted"[\s\S]*startSentryInitOnce\("app-mounted"\)/, 'Sentry init module must register the app-mounted Sentry init trigger inside attachSentry');
+assert.match(frontendMain, /import \{ attachSentry, captureRenderError \} from "@\/lib\/sentry\/init"/, 'Frontend main must consume Sentry attach + capture from the extracted module');
+assertOrder(frontendMain, 'attachSentry();', 'renderApp();', 'Frontend boot must call attachSentry before rendering so the app-mounted Sentry trigger is registered');
+assertOrder(frontendMain, 'window.addEventListener(\n  "nixmac:app-mounted"', 'renderApp();', 'Frontend boot must register the heartbeat-stop app-mounted listener before rendering the app');
 assert.doesNotMatch(frontendMain, /renderApp\(\);\s*void initializeSentry/, 'Frontend boot must not directly initialize preference-backed Sentry immediately after first render');
 assert.doesNotMatch(frontendRenderApp, /\bawait\b/, 'Frontend renderApp must stay synchronous and never await prefs IPC before first render');
-assert.match(frontendMain, /role="alert"[\s\S]*background: "#27272a"[\s\S]*border: "1px solid #52525b"/, 'Frontend error fallback must include a visible central card with enough luminance range for screenshot signal diagnostics');
+assert.match(frontendStartupFallback, /role="alert"[\s\S]*background: "#27272a"[\s\S]*border: "1px solid #52525b"/, 'Startup fallback must include a visible central card with enough luminance range for screenshot signal diagnostics');
 assert.match(frontendMain, /window\.addEventListener\("unhandledrejection"[\s\S]*window unhandled rejection/, 'Frontend boot diagnostics must capture top-level unhandled rejections');
 assert.match(runnerShell, /E2E_TERMINAL_CLEANUP_MODE=kill recording_close_terminal_windows/, 'Runner preflight must kill stale recorder Terminal windows before each scenario');
 assert.match(peekabooRunner, /for key in NIXMAC_E2E_MOCK_SYSTEM NIXMAC_E2E_SOLID_CAPTURE NIXMAC_E2E_OPAQUE_WINDOW NIXMAC_E2E_WEBVIEW_WATCHDOG NIXMAC_SKIP_PERMISSIONS/, 'Runner preflight must clear stale Peekaboo launchctl flags, including solid capture, opaque capture, and the independent WebView watchdog');
